requirements.txt
flask==3.0.3
moviepy==1.0.3
openai-whisper
